pyproject.toml
[tool.poetry]
name = "mnist_experiments"
version = "0.1.0"
description = "Learning experiments with DL and MNIST data"
authors = ["kolommik <mikhail.kolomasov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7.2, <3.11"
tensorflow-gpu = "^2.9.1"
matplotlib = "^3.5.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
mypy = "^0.971"
pylint = "^2.14.5"
ipykernel = "^6.15.1"
tensorboard = "^2.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
